Refactor :: S3 파일 이동
